Added blackjack command.
